+++ .eslintrc.json
@@ -82,6 +82,11 @@
             "position": "before"
           },
           {
+            "pattern": "@/services/**",
+            "group": "internal",
+            "position": "before"
+          },
+          {
             "pattern": "@/hooks/**",
             "group": "internal",
             "position": "before"
