+++ tsconfig.json
@@ -26,6 +26,7 @@
       "@/content/*": ["./src/content/*"],
       "@/data/*": ["./src/app/_data/*"],
       "@/hooks/*": ["./src/app/_hooks/*"],
+      "@/services/*": ["./src/app/_services/*"],
       "@/styles/*": ["./src/app/_styles/*"],
       "@/types/*": ["./src/app/_types/*"],
       "@/schemas/*": ["./src/app/_schemas/*"],
@@ -41,8 +42,7 @@
     ".next/types/**/*.ts",
     "src/app/_types/**/*.d.ts",
     "src/app/events/[slug]",
-    "cypress/**/*.ts",
-    "src/app/ecosystem-explorer/project-form/services/github/config"
+    "cypress/**/*.ts"
   ],
   "exclude": ["node_modules"]
 }
